user_crud/requirements.txt
Flask
Flask_RESTful
flask_sqlalchemy
SQLAlchemy
cryptography
pymysql
python-dotenv
Flask-JWT-Extended
redis